node:tls: move the transport error fixture to a file, run it with bunRun
Also shorten the comment on the 'error' forward in attachUpgradedDuplex.
